Calificacion y archivos de prueba
Calificación y archivos de prueba

+++ Analizador/Interprete/Tabla_de_Simbolos/Tipo.py
@@ -2,7 +2,7 @@ from enum import Enum
 
 class TIPO(Enum):
     ENTERO   = 1
-    DECIMAL  = 2
+    FLOAT  = 2
     BOOLEANO = 3
     CARACTER = 4
     CADENA   = 5
